Remove dependency on openssl-windows
OpenSSL is included with ruby-windows, so there is already
a verison available for chef to use. If projects want a version
different than that provided by ruby-windows, then that should
be defined at the project level.

The reason it is being removed now is because there is a ruby
that has an openssl I do not want clobbered.